// src-old/proto/js/Agent_pb.d.ts
export class EmptyMessage extends jspb.Message { 

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): EmptyMessage.AsObject;
    static toObject(includeInstance: boolean, msg: EmptyMessage): EmptyMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: EmptyMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): EmptyMessage;
    static deserializeBinaryFromReader(message: EmptyMessage, reader: jspb.BinaryReader): EmptyMessage;
}

export namespace EmptyMessage {
    export type AsObject = {
    }
}
// ...
export class AgentStatusMessage extends jspb.Message { 
    getStatus(): AgentStatusType;
    setStatus(value: AgentStatusType): AgentStatusMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): AgentStatusMessage.AsObject;
    static toObject(includeInstance: boolean, msg: AgentStatusMessage): AgentStatusMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: AgentStatusMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): AgentStatusMessage;
    static deserializeBinaryFromReader(message: AgentStatusMessage, reader: jspb.BinaryReader): AgentStatusMessage;
}

export namespace AgentStatusMessage {
    export type AsObject = {
        status: AgentStatusType,
    }
}

export class NewKeyPairMessage extends jspb.Message { 
    getPassphrase(): string;
    setPassphrase(value: string): NewKeyPairMessage;

    getNbits(): number;
    setNbits(value: number): NewKeyPairMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): NewKeyPairMessage.AsObject;
    static toObject(includeInstance: boolean, msg: NewKeyPairMessage): NewKeyPairMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: NewKeyPairMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): NewKeyPairMessage;
    static deserializeBinaryFromReader(message: NewKeyPairMessage, reader: jspb.BinaryReader): NewKeyPairMessage;
}

export namespace NewKeyPairMessage {
    export type AsObject = {
        passphrase: string,
        nbits: number,
    }
}
// ...
export class UnlockNodeMessage extends jspb.Message { 
    getPassphrase(): string;
    setPassphrase(value: string): UnlockNodeMessage;

    getTimeout(): number;
    setTimeout(value: number): UnlockNodeMessage;


    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): UnlockNodeMessage.AsObject;
    static toObject(includeInstance: boolean, msg: UnlockNodeMessage): UnlockNodeMessage.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: UnlockNodeMessage, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): UnlockNodeMessage;
    static deserializeBinaryFromReader(message: UnlockNodeMessage, reader: jspb.BinaryReader): UnlockNodeMessage;
}

export namespace UnlockNodeMessage {
    export type AsObject = {
        passphrase: string,
        timeout: number,
    }
}
// ...
export enum AgentStatusType {
    ONLINE = 0,
    OFFLINE = 1,
    ERRORED = 2,
}